try to fix the routing
