+++ package.json
@@ -10,8 +10,6 @@
     "test": "jest --watchAll"
   },
   "dependencies": {
-    "cloudinary-react": "^1.8.1",
-    "date-fns": "^2.30.0",
     "eslint": "8.45.0",
     "eslint-config-next": "13.4.10",
     "mongodb": "^5.7.0",
