+++ pages/index.js
@@ -1,7 +1,7 @@
 import Head from "next/head";
 import styled from "styled-components";
 import { Inter } from "next/font/google";
-import KidsList from "@/components/KidsList/index";
+import KidsList from "@/components/KidsList";
 
 const inter = Inter({ subsets: ["latin"] });
